Added Slice of the displayNewsletter
